Python script fixes

+++ bin/virtualenv.bash
@@ -22,7 +22,7 @@ create_virtualenv() {
     fi
 
     "$VIRTUAL_ENV/bin/$PYTHON" -m pip install $PACKAGE $DEPENDS || exit 1
-    find "$VIRTUAL_ENV/lib" -type f -name '*test*.py' | grep "/[^/]*test[^/]*/" | sed -e "s/\/[^\/]*$//" | uniq | xargs rm -rfv
+    [ "${1:-}" != "-noclean" ] && find "$VIRTUAL_ENV/lib" -type f -name '*test*.py' | grep "/[^/]*test[^/]*/" | sed -e "s/\/[^\/]*$//" | uniq | xargs rm -rfv
     IFS=$'\n'
     for FILE in $(grep "^#!/.*/python" "$VIRTUAL_ENV/bin"/* 2> /dev/null | grep -v ":#!/usr/bin/env " | sed -e "s@:#!/.*@@")
     do
@@ -39,8 +39,9 @@ create_virtualenv() {
 PYTHON_DIR=$(echo "import sys; print(sys.exec_prefix)" | "$PYTHON")
 VIRTUAL_ENV="$PYTHON_DIR-venv/${PACKAGE/==/-}"
 [ -d "$VIRTUAL_ENV" ] || [ -w "${VIRTUAL_ENV%/*/*}" ] || VIRTUAL_ENV="$TMP/$($PYTHON --version 2>&1 | sed -e "s/ /-/g")-venv/${PACKAGE/==/-}"
+FLAGS="${1:-}"
 export VIRTUAL_ENV
 export PATH="$VIRTUAL_ENV/bin:$PATH"
-[ "$("$VIRTUAL_ENV/bin/$PYTHON" -m pip freeze 2> /dev/null | grep -i "^$PACKAGE$")" ] || create_virtualenv
+[ "$("$VIRTUAL_ENV/bin/$PYTHON" -m pip freeze 2> /dev/null | grep -i "^$PACKAGE$")" ] || create_virtualenv "$@"
 
 [ "${0##*-}" = venv ] && exec bash -l
